refactor: 유틸 패키지 생성 및 FixedPageRequest 분리
