+++ src/main/kotlin/com/wafflestudio/csereal/core/notice/service/NoticeService.kt
@@ -1,7 +1,7 @@
 package com.wafflestudio.csereal.core.notice.service
 
 import com.wafflestudio.csereal.common.CserealException
-import com.wafflestudio.csereal.common.cleanTextFromHtml
+import com.wafflestudio.csereal.common.utils.cleanTextFromHtml
 import com.wafflestudio.csereal.core.notice.database.*
 import com.wafflestudio.csereal.core.notice.dto.*
 import com.wafflestudio.csereal.core.resource.attachment.service.AttachmentService
